# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Api Test",
        "description": "Api Swagger pour le test",
        "contact": {
            "email": "tshipambavincent80@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API Server TEST"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Enregistrement d'un nouvel utilisateur"
                ],
                "summary": "Créer un compte utilisateur",
                "operationId": "5d4fa5d5607ed675ba93907e1fc94534",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nom d'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Adresse mail",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Connexion réussie"
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Déconnexion"
                ],
                "summary": "Déconnexion d'un compte utilisateur",
                "operationId": "28d48c13fca984400ac8de99f54ee13a",
                "responses": {
                    "200": {
                        "description": "Utilisateur déconnecté avec succès !"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Afficher la liste des utilisateurs",
                "description": "Api qui nous retourne la liste des utilisateurs",
                "operationId": "getUsersList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Enregistrement d'un nouvel utilisateur",
            "description": "Enregistrement d'un nouvel utilisateur"
        },
        {
            "name": "Authentification",
            "description": "Authentification"
        },
        {
            "name": "Déconnexion",
            "description": "Déconnexion"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}